Confirmando a exclusão e exibindo um Toast

+++ src/components/TaskListTable.js
@@ -1,5 +1,7 @@
 import React, { Component } from 'react';
 import TaskService from '../api/TaskService';
+import { ToastContainer, toast } from 'react-toastify';
+import "react-toastify/dist/ReactToastify.css";
 
 class TaskListTable extends Component {
     constructor(props) {
@@ -21,19 +23,25 @@ class TaskListTable extends Component {
     }
 
     onDeleteHandler(id) {
-        TaskService.delete(id);
-        this.listTasks();
+        if (window.confirm("Deseja realmente excluit essa tarefa?")) {
+            TaskService.delete(id);
+            this.listTasks(); 
+            toast.success("Tarefa excluída com sucesso!", {position: toast.POSITION.TOP_CENTER})
+        };
     }
 
     render() {
         return (
-            <table className="table table-striped">
-                <TableHeader/>
-                <TableBody 
-                    tasks={this.state.tasks}
-                    onDelete={this.onDeleteHandler}
-                />
-            </table>
+            <>
+                <table className="table table-striped">
+                    <TableHeader/>
+                    <TableBody 
+                        tasks={this.state.tasks}
+                        onDelete={this.onDeleteHandler}
+                    />
+                </table>
+                <ToastContainer autoClose={3000}/>
+            </>
         );
     }
